+++ src/components/User/RegisterLocation/LocationClickResult.tsx
@@ -1,23 +1,16 @@
 import { Dispatch, SetStateAction } from "react";
+import styled from "styled-components";
 import { getLocationName } from "../RegisterLocation";
 import { RoundButton, TargetButton } from "components/Common/Buttons";
-import { useHistory } from "react-router-dom";
-import { ROUTE } from "util/constants";
-import styled from "styled-components";
+
 interface LocationClickResultProps {
   isSameDong: boolean;
   searchLogs: string[];
   setSearchLogs: Dispatch<SetStateAction<string[]>>;
 }
 export default function LocationClickResult({ isSameDong, searchLogs, setSearchLogs }: LocationClickResultProps) {
-  const history = useHistory();
   const handleDeleteTarget = (location: string) => () => {
     setSearchLogs((prev: string[]) => prev.filter((target) => target !== location));
-  };
-
-  const handleNextPage = () => {
-    if (searchLogs.length !== 2) return;
-    history.push(ROUTE.MAIN);
   };
 
   return (
@@ -36,10 +29,6 @@ export default function LocationClickResult({ isSameDong, searchLogs, setSearchL
           );
         })}
       </ButtonBox>
-
-      <NextButton onClick={handleNextPage} variant="outlined">
-        다음 ({searchLogs.length}/2)
-      </NextButton>
     </LocationBox>
   );
 }
